+++ src/main/java/com/ravi/orbit/dto/SizeDTO.java
@@ -10,12 +10,14 @@ import jakarta.persistence.ManyToOne;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class SizeDTO {
 
-    public SizeDTO(Long id, String size, boolean isAvailable, Double price, Integer quantity, Long productId) {
+    public SizeDTO(Long id, String size, boolean isAvailable, BigDecimal price, Integer quantity, Long productId) {
         this.id = id;
         this.size = size;
         this.isAvailable = isAvailable;
@@ -30,7 +32,7 @@ public class SizeDTO {
 
     private boolean isAvailable;
 
-    private Double price;
+    private BigDecimal price;
 
     private Integer quantity;
